requirements.txt
streamlit
pandas
azure-ai-inference
python-dotenv
openpyxl
plotly
PyPDF2